Major refactor almost complete - using ComfyUI model management and patching mechanisms to do injection/ejection, consolidate AD-type motion module code into one
