+++ animatediff/nodes_experimental.py
@@ -1,7 +1,7 @@
 
 import torch
 
-from .motion_module import MotionModelSettings
+from .model_injection import MotionModelSettings
 
 
 class AnimateDiffModelSettingsSimple:
